♻️ [Refactoring]: #041 core/src/document/ 直下を 1機能1フォルダ化
- catalog-parser, decode-pdf-string, document-info-parser, document-metadata, pdf-date, pdf-doc-encoding, pdf-document, pdf-page の 8 件を {feature}/index.ts へフォルダ化
- 各 test / helpers を {feature}/__tests__/ へ git mv
- 主実装内の相対 import を +1、test 内を +2 階層調整
- helper 参照は同一 __tests__/ folder 内なので ./{feat}.test.helpers へ復元
- document/index.ts (barrel) は folder/index 解決で無修正
- Test Files=124 / Tests=1688 維持
