+++ packages/core/src/document/document-metadata/__tests__/document-metadata.error.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "vitest";
-import type { PdfWarning } from "../pdf/errors/warning/index";
-import type { PdfValue } from "../pdf/types/pdf-types/index";
-import { parseTrappedName } from "./document-metadata";
+import type { PdfWarning } from "../../../pdf/errors/warning/index";
+import type { PdfValue } from "../../../pdf/types/pdf-types/index";
+import { parseTrappedName } from "../../document-metadata";
 
 const makeName = (value: string): PdfValue => ({ type: "name", value });
