+++ packages/core/src/document/document-metadata/__tests__/document-metadata.validation.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { PdfTrapped } from "./document-metadata";
+import { PdfTrapped } from "../../document-metadata";
 
 test.each([
   ["True"],
