+++ packages/core/src/document/pdf-date/__tests__/pdf-date.error.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { parsePdfDate } from "./pdf-date";
+import { parsePdfDate } from "../../pdf-date";
 
 test.each([
   ["20230101"],
