+++ packages/core/src/document/pdf-date/__tests__/pdf-date.parse.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { parsePdfDate } from "./pdf-date";
+import { parsePdfDate } from "../../pdf-date";
 
 test("D: プレフィックスを欠いた文字列は undefined を返す", () => {
   expect(parsePdfDate("20230101")).toBeUndefined();
