+++ packages/core/src/document/pdf-date/index.ts
