+++ packages/core/src/document/pdf-document/__tests__/pdf-document.boundary.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test } from "vitest";
-import { PdfDocument } from "./pdf-document";
+import { PdfDocument } from "../../pdf-document";
 import { buildMinimalSinglePagePdf } from "./pdf-document.test.helpers";
 
 test.each([
